Fix build warnings.

+++ tlsBIO.c
@@ -1,7 +1,7 @@
 /*
  * Copyright (C) 1997-2000 Matt Newman <matt@novadigm.com>
  *
- * $Header: /cvsroot/tls/tls/tlsBIO.c,v 1.6 2002/02/04 22:46:31 hobbs Exp $
+ * $Header: /cvsroot/tls/tls/tlsBIO.c,v 1.7 2004/02/04 04:02:19 razzell Exp $
  *
  * Provides BIO layer to interface openssl to Tcl.
  */
@@ -129,7 +129,7 @@ BioPuts	(bio, str)
     BIO *bio;
     CONST char *str;
 {
-    return BioWrite(bio, str, strlen(str));
+    return BioWrite(bio, str, (int) strlen(str));
 }
 
 static long
