+++ tlsInt.h
@@ -1,7 +1,7 @@
 /*
  * Copyright (C) 1997-2000 Matt Newman <matt@novadigm.com>
  *
- * $Header: /cvsroot/tls/tls/tlsInt.h,v 1.10 2003/05/15 21:02:10 razzell Exp $
+ * $Header: /cvsroot/tls/tls/tlsInt.h,v 1.11 2004/02/04 04:02:19 razzell Exp $
  *
  * TLS (aka SSL) Channel - can be layered on any bi-directional
  * Tcl_Channel (Note: Requires Trf Core Patch)
@@ -66,7 +66,7 @@
 #endif
 
 #define SSL_ERROR(ssl,err)	\
-	    ((char*)ERR_reason_error_string(SSL_get_error((ssl),(err))))
+    ((char*)ERR_reason_error_string((unsigned long)SSL_get_error((ssl),(err))))
 /*
  * OpenSSL BIO Routines
  */
